clean-up Lua
Change the way the Lua formatting script is handled.  Instead of a
separate file that needs to live somewhere relative to the working
directory of the bot the format script is now just part of the
configuration.  It can be given as an in-line value in YAML like so:

  bot:
    msg_expiration: 4
    msg_capacity: 100
    format_script:
      function main(event)
          return "there was an event"
      end

The "format_script" key is optional however and if it isn't present the
value is taken from the scripts/format.Lua file.  That file however is
baked in at compile time.  There is now a syntax check that will fail
the bot startup if the Lua cannot be parsed or the main function is
missing.  On any other error detected at run-time panics in the
formatting will now produce an error message instead of a panic.

Fixes #46