make gerrit::Event into enum
gerrit::Event is a struct with an event_type field and many fields being
modeled as option.  This is cumbersome and error prone because to use it
one would have to make assumptions or check the documentation about
which fields are being present when.  In actuality the value of the
event_type field determines which fields should be present and this is
well documented in Gerrit's documentation. Commit makes gerrit::Event
into an enum with (currently) two variants: CommentAdded and
ReviewerAdded.  Both are supported by their own structs.  This makes
accessing common fields more difficult but improves the logic in the bot
and (hopefully) simplifies treating new event types in the future.